requirements.txt
beautifulsoup4==4.11.1
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorgram.py==1.2.0
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
flash==1.0.3
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
PyJWT==2.4.0
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
twilio==7.14.0
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
pymysql
gunicorn
psycopg2